+++ src/apps/ocioview/ocioview/viewer/__init__.py
@@ -1,8 +1,6 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright Contributors to the OpenColorIO Project.
 
-from .base_image_viewer import ViewerChannels, BaseImageViewer
-from .edit_image_viewer import EditImageViewer
+from .image_viewer import ViewerChannels, ImageViewer
 from .offscreen_viewer import WgpuCanvasOffScreenViewer
-from .preview_image_viewer import PreviewImageViewer
 from .utils import load_image
