+++ src/apps/ocioview/ocioview/viewer/image_viewer.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 
 from contextlib import contextmanager
 from pathlib import Path
-from typing import Generator, Optional
+from typing import Generator, Optional, Union
 
 import PyOpenColorIO as ocio
 from PySide6 import QtCore, QtGui, QtWidgets
@@ -18,9 +18,15 @@ from ..constants import (
     B_COLOR,
     ICON_SIZE_TAB,
 )
+from ..items.display_model import DisplayModel
+from ..items.view_model import ViewModel
+from ..mode import OCIOViewMode
 from ..processor_context import ProcessorContext
+from ..ref_space_manager import ReferenceSpaceManager
+from ..signal_router import SignalRouter
+from ..transform_manager import TransformManager
 from ..utils import float_to_uint8, get_glyph_icon, SignalsBlocked
-from ..widgets import CallbackComboBox, ColorSpaceComboBox
+from ..widgets import ComboBox, CallbackComboBox, ColorSpaceComboBox
 from .image_plane import ImagePlane
 
 
@@ -33,9 +39,9 @@ class ViewerChannels(object):
     R, G, B, A, ALL = list(range(5))
 
 
-class BaseImageViewer(QtWidgets.QWidget):
+class ImageViewer(QtWidgets.QWidget):
     """
-    Base image viewer widget, which can display an image with internal
+    Image viewer widget, which can display an image with internal
     32-bit float precision.
     """
 
@@ -45,6 +51,13 @@ class BaseImageViewer(QtWidgets.QWidget):
     FMT_B_LABEL = f'<span style="color:{B_COLOR.name()};">{{v}}</span>'
     FMT_SWATCH_CSS = "background-color: rgb({r}, {g}, {b});"
     FMT_IMAGE_SCALE = f'{{s:,d}}{FMT_GRAY_LABEL.format(v="%")}'
+
+    PASSTHROUGH = "passthrough"
+    PASSTHROUGH_LABEL = FMT_GRAY_LABEL.format(v=f"{PASSTHROUGH}:")
+
+    ROLE_SLOT = QtCore.Qt.UserRole + 1
+    ROLE_ITEM_TYPE = QtCore.Qt.UserRole + 2
+    ROLE_ITEM_NAME = QtCore.Qt.UserRole + 3
 
     @classmethod
     def viewer_type_icon(cls) -> QtGui.QIcon:
@@ -60,8 +73,14 @@ class BaseImageViewer(QtWidgets.QWidget):
         super().__init__(parent)
 
         self._sample_format = ""
+        self._tf_subscription_slot = -1
+        self._tf_fwd = None
+        self._tf_inv = None
 
         # Widgets
+        # ---------------------------------------------------------------------
+
+        # Viewport
         self.image_plane = ImagePlane(self)
         self.image_plane.setSizePolicy(
             QtWidgets.QSizePolicy(
@@ -70,6 +89,7 @@ class BaseImageViewer(QtWidgets.QWidget):
             )
         )
 
+        # Input color space
         self.input_color_space_label = get_glyph_icon(
             "mdi6.import", as_widget=True
         )
@@ -79,6 +99,51 @@ class BaseImageViewer(QtWidgets.QWidget):
             self.input_color_space_label.toolTip()
         )
 
+        # Edit mode
+        self.tf_label = get_glyph_icon("mdi6.export", as_widget=True)
+        self.tf_box = ComboBox()
+        self.tf_box.setToolTip("Output transform")
+
+        self._tf_direction_forward_icon = get_glyph_icon("mdi6.layers-plus")
+        self._tf_direction_inverse_icon = get_glyph_icon("mdi6.layers-minus")
+
+        self.tf_direction_button = QtWidgets.QPushButton()
+        self.tf_direction_button.setCheckable(True)
+        self.tf_direction_button.setChecked(False)
+        self.tf_direction_button.setIcon(self._tf_direction_forward_icon)
+        self.tf_direction_button.setToolTip("Transform direction: Forward")
+
+        self.output_tf_direction_label = QtWidgets.QLabel("+")
+
+        # Preview mode
+        self.display_view_label = get_glyph_icon(
+            ViewModel.__icon_glyph__, as_widget=True
+        )
+        self.display_view_label.setToolTip(
+            f"{DisplayModel.item_type_label()} / {ViewModel.item_type_label()}"
+        )
+
+        self.display_box = CallbackComboBox(
+            self._get_displays,
+            get_default_item=self._get_default_display,
+        )
+        self.display_box.setSizeAdjustPolicy(
+            QtWidgets.QComboBox.AdjustToContents
+        )
+        self.display_box.setToolTip(DisplayModel.item_type_label())
+        self.display_box.currentIndexChanged[int].connect(
+            self._on_display_changed
+        )
+
+        self.view_box = CallbackComboBox(
+            self._get_views,
+            get_default_item=self._get_default_view,
+        )
+        self.view_box.setSizeAdjustPolicy(QtWidgets.QComboBox.AdjustToContents)
+        self.view_box.setToolTip(ViewModel.item_type_label())
+        self.view_box.currentIndexChanged[int].connect(self._on_view_changed)
+
+        # Image adjustments
         self.exposure_label = get_glyph_icon("ph.aperture", as_widget=True)
         self.exposure_label.setToolTip("Exposure")
         self.exposure_box = QtWidgets.QDoubleSpinBox()
@@ -106,6 +171,7 @@ class BaseImageViewer(QtWidgets.QWidget):
         )
         self.sample_precision_box.setValue(5)
 
+        # Info and inspect labels
         self.image_name_label = QtWidgets.QLabel()
         self.image_scale_label = QtWidgets.QLabel(
             self.FMT_IMAGE_SCALE.format(s=100)
@@ -153,6 +219,9 @@ class BaseImageViewer(QtWidgets.QWidget):
         )
 
         # Layout
+        # ---------------------------------------------------------------------
+
+        # Info and inspect labels
         self.info_layout = QtWidgets.QHBoxLayout()
         self.info_layout.setContentsMargins(8, 8, 8, 8)
         self.info_layout.addWidget(self.image_name_label)
@@ -235,11 +304,59 @@ class BaseImageViewer(QtWidgets.QWidget):
         )
         self.inspect_bar.setLayout(self.inspect_layout)
 
+        # Edit mode
+        self.tf_layout = QtWidgets.QHBoxLayout()
+        self.tf_layout.setContentsMargins(0, 0, 0, 0)
+        self.tf_layout.setSpacing(0)
+        self.tf_layout.addWidget(self.tf_box)
+        self.tf_layout.setStretch(0, 1)
+        self.tf_layout.addWidget(self.tf_direction_button)
+
+        self.tf_layout_outer = QtWidgets.QHBoxLayout()
+        self.tf_layout_outer.setContentsMargins(0, 0, 0, 0)
+        self.tf_layout_outer.addWidget(self.tf_label)
+        self.tf_layout_outer.setStretch(0, 0)
+        self.tf_layout_outer.addLayout(self.tf_layout)
+        self.tf_layout_outer.setStretch(1, 1)
+
+        self.tf_frame = QtWidgets.QFrame()
+        self.tf_frame.setLayout(self.tf_layout_outer)
+
+        self.inspect_layout.addWidget(
+            self.output_tf_direction_label, 1, 5, QtCore.Qt.AlignRight
+        )
+
+        # Preview mode
+        self.display_view_layout = QtWidgets.QHBoxLayout()
+        self.display_view_layout.setContentsMargins(0, 0, 0, 0)
+        self.display_view_layout.setSpacing(0)
+        self.display_view_layout.addWidget(self.display_box)
+        self.display_view_layout.addWidget(self.view_box)
+
+        self.display_view_layout_outer = QtWidgets.QHBoxLayout()
+        self.display_view_layout_outer.setContentsMargins(0, 0, 0, 0)
+        self.display_view_layout_outer.addWidget(self.display_view_label)
+        self.display_view_layout_outer.setStretch(0, 0)
+        self.display_view_layout_outer.addLayout(self.display_view_layout)
+        self.display_view_layout_outer.setStretch(1, 1)
+
+        self.display_view_frame = QtWidgets.QFrame()
+        self.display_view_frame.setLayout(self.display_view_layout_outer)
+
+        # Mode switch stack
+        self.mode_stack = QtWidgets.QStackedWidget()
+        self.mode_stack.addWidget(self.tf_frame)  # Edit mode
+        self.mode_stack.addWidget(self.display_view_frame)  # Preview mode
+
+        # Input/output
         self.io_layout = QtWidgets.QHBoxLayout()
         self.io_layout.addWidget(self.input_color_space_label)
         self.io_layout.addWidget(self.input_color_space_box)
         self.io_layout.setStretch(1, 1)
+        self.io_layout.addWidget(self.mode_stack)
+        self.io_layout.setStretch(2, 1)
 
+        # Image adjustments
         self.adjust_layout = QtWidgets.QHBoxLayout()
         self.adjust_layout.addWidget(self.exposure_label)
         self.adjust_layout.addWidget(self.exposure_box)
@@ -251,19 +368,24 @@ class BaseImageViewer(QtWidgets.QWidget):
         self.adjust_layout.addWidget(self.sample_precision_box)
         self.adjust_layout.setStretch(5, 1)
 
+        # Viewport
         self.image_plane_layout = QtWidgets.QVBoxLayout()
         self.image_plane_layout.setSpacing(0)
         self.image_plane_layout.addWidget(self.info_bar)
         self.image_plane_layout.addWidget(self.image_plane)
         self.image_plane_layout.addWidget(self.inspect_bar)
 
+        # Main layout
         layout = QtWidgets.QVBoxLayout()
         layout.addLayout(self.io_layout)
         layout.addLayout(self.adjust_layout)
         layout.addLayout(self.image_plane_layout)
+        layout.setStretch(2, 1)
         self.setLayout(layout)
 
         # Connect signals/slots
+        # ---------------------------------------------------------------------
+
         self.image_plane.image_loaded.connect(self._on_image_loaded)
         self.image_plane.scale_changed.connect(self._on_scale_changed)
         self.image_plane.sample_changed.connect(self._on_sample_changed)
@@ -276,9 +398,36 @@ class BaseImageViewer(QtWidgets.QWidget):
             self._on_sample_precision_changed
         )
 
+        signal_router = SignalRouter.get_instance()
+        signal_router.mode_changed.connect(self._on_mode_changed)
+
+        # Edit mode
+        self.image_plane.tf_subscription_requested.connect(
+            self._on_tf_subscription_requested
+        )
+        self.tf_box.currentIndexChanged[int].connect(
+            self._on_transform_changed
+        )
+        self.tf_direction_button.clicked[bool].connect(
+            self._on_inverse_check_clicked
+        )
+
         # Initialize
+        # ---------------------------------------------------------------------
+
         self._on_sample_precision_changed(self.sample_precision_box.value())
         self._on_sample_changed(-1, -1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0)
+
+        # Edit mode
+        TransformManager.subscribe_to_transform_menu(
+            self._on_transform_menu_changed
+        )
+        TransformManager.subscribe_to_transform_subscription_init(
+            self._on_transform_subscription_init
+        )
+
+        # Initialize viewport
+        self.update(force=True)
 
     def update(self, force: bool = False, update_items: bool = True) -> None:
         """
@@ -290,6 +439,17 @@ class BaseImageViewer(QtWidgets.QWidget):
         :param update_items: Whether to update dynamic OCIO items that
             affect image processing.
         """
+        mode = OCIOViewMode.current_mode()
+        if mode == OCIOViewMode.Preview and update_items:
+            self.display_box.update_items()
+            self.view_box.update_items()
+
+        self.image_plane.update_ocio_proc(
+            proc_context=self._make_processor_context(),
+            transform=self._make_transform(),
+            force_update=force,
+        )
+
         super().update()
 
         # Broadcast this viewer's image data for other app components
@@ -300,9 +460,12 @@ class BaseImageViewer(QtWidgets.QWidget):
         self.image_plane.reset_ocio_proc(update=False)
 
         # Update widgets to match image plane
-        with SignalsBlocked(self):
+        with self._ocio_signals_blocked():
+            self.set_transform_direction(ocio.TRANSFORM_DIR_FORWARD)
             self.set_exposure(self.image_plane.exposure())
             self.set_gamma(self.image_plane.gamma())
+            self.display_box.reset()
+            self.view_box.reset()
 
         # Update input color spaces and redraw viewport
         self.update()
@@ -373,6 +536,92 @@ class BaseImageViewer(QtWidgets.QWidget):
         """
         self.gamma_box.setValue(value)
 
+    def display(self) -> str:
+        """Get current OCIO display."""
+        return self.display_box.currentText()
+
+    def view(self) -> str:
+        """Get current OCIO view."""
+        return self.view_box.currentText()
+
+    def transform(self) -> Optional[ocio.Transform]:
+        """Get current OCIO transform."""
+        return self.image_plane.transform()
+
+    def set_transform(
+        self,
+        slot: int,
+        transform_fwd: Optional[ocio.Transform],
+        transform_inv: Optional[ocio.Transform],
+    ) -> None:
+        """
+        Update main OCIO transform for the viewing pipeline, to be
+        applied from the current config's scene reference space.
+
+        :param slot: Transform subscription slot
+        :param transform_fwd: Forward transform
+        :param transform_inv: Inverse transform
+        """
+        tf_direction = self.transform_direction()
+
+        if (
+            slot != self._tf_subscription_slot
+            or (
+                transform_fwd is None
+                and tf_direction == ocio.TRANSFORM_DIR_FORWARD
+            )
+            or (
+                transform_inv is None
+                and tf_direction == ocio.TRANSFORM_DIR_INVERSE
+            )
+        ):
+            return
+
+        self._tf_fwd = transform_fwd
+        self._tf_inv = transform_inv
+
+        self.update()
+
+    def clear_transform(self) -> None:
+        """
+        Clear current OCIO transform, passing through the input image.
+        """
+        self._tf_subscription_slot = -1
+        self._tf_fwd = None
+        self._tf_inv = None
+
+        if self.tf_box.currentIndex() != 0:
+            with SignalsBlocked(self.tf_box):
+                self.tf_box.setCurrentIndex(0)
+
+        self.image_plane.clear_transform()
+
+    def transform_item_type(self) -> type | None:
+        """Get transform source config item type."""
+        return self.tf_box.currentData(role=self.ROLE_ITEM_TYPE)
+
+    def transform_item_name(self) -> str | None:
+        """Get transform source config item name."""
+        return self.tf_box.currentData(role=self.ROLE_ITEM_NAME)
+
+    def transform_direction(self) -> ocio.TransformDirection:
+        """Get transform direction being viewed."""
+        return (
+            ocio.TRANSFORM_DIR_INVERSE
+            if self.tf_direction_button.isChecked()
+            else ocio.TRANSFORM_DIR_FORWARD
+        )
+
+    def set_transform_direction(
+        self, direction: ocio.TransformDirection
+    ) -> None:
+        """
+        :param direction: Set the transform direction to be viewed
+        """
+        self.tf_direction_button.setChecked(
+            direction == ocio.TRANSFORM_DIR_INVERSE
+        )
+
     @contextmanager
     def _ocio_signals_blocked(self) -> Generator:
         """
@@ -383,29 +632,74 @@ class BaseImageViewer(QtWidgets.QWidget):
         self.input_color_space_box.blockSignals(True)
         self.exposure_box.blockSignals(True)
         self.gamma_box.blockSignals(True)
+        self.display_box.blockSignals(True)
+        self.view_box.blockSignals(True)
 
         yield
 
         self.input_color_space_box.blockSignals(False)
         self.exposure_box.blockSignals(False)
         self.gamma_box.blockSignals(False)
+        self.display_box.blockSignals(False)
+        self.view_box.blockSignals(False)
 
     def _make_processor_context(self) -> ProcessorContext:
         """Create processor context from available data."""
-        return ProcessorContext(self.input_color_space())
+        mode = OCIOViewMode.current_mode()
+        if mode == OCIOViewMode.Preview:
+            return ProcessorContext(
+                self.input_color_space(),
+                ViewModel.__item_type__,
+                self.view(),
+                ocio.TRANSFORM_DIR_FORWARD,
+            )
+        else:  # Edit
+            return ProcessorContext(
+                self.input_color_space(),
+                self.transform_item_type(),
+                self.transform_item_name(),
+                self.transform_direction(),
+            )
 
-    def _get_color_spaces(self) -> list[str]:
-        """
-        Get all active color spaces defined relative to the config's
-        scene reference space.
-        """
-        return ConfigCache.get_color_space_names(
-            ocio.SEARCH_REFERENCE_SPACE_SCENE
-        )
+    def _make_transform(self) -> Union[ocio.Transform, None]:
+        """Create viewer transform."""
+        transform = None
+        mode = OCIOViewMode.current_mode()
+
+        if mode == OCIOViewMode.Preview:
+            display = self.display()
+            view = self.view()
+
+            if display and view:
+                # Image plane expects all transforms to be relative to the current
+                # config's scene reference space.
+                transform = ocio.DisplayViewTransform(
+                    src=ReferenceSpaceManager.scene_reference_space().getName(),
+                    display=display,
+                    view=view,
+                    direction=ocio.TRANSFORM_DIR_FORWARD,
+                )
+
+        else:  # Edit
+            if self._tf_fwd is not None and self._tf_inv is not None:
+                if self.transform_direction() == ocio.TRANSFORM_DIR_INVERSE:
+                    return self._tf_inv
+                else:
+                    return self._tf_fwd
+            else:
+                # Return no-op transform. Returning None instead results in the image
+                # plane processor being unchanged, which is problematic when switching
+                # application modes, since it will retain the previous display/view
+                # transform.
+                return ocio.ExponentTransform()
+
+        return transform
 
     def _get_default_color_space(self) -> str:
         """Get reasonable default color space."""
-        all_color_spaces = self._get_color_spaces()
+        all_color_spaces = ConfigCache.get_color_space_names(
+            ocio.SEARCH_REFERENCE_SPACE_SCENE
+        )
         default_color_space = ConfigCache.get_default_color_space_name()
         if (
             default_color_space is not None
@@ -416,6 +710,51 @@ class BaseImageViewer(QtWidgets.QWidget):
             return all_color_spaces[0]
         else:
             return ""
+
+    def _get_displays(self) -> list[str]:
+        """Get all active OCIO displays."""
+        config = ocio.GetCurrentConfig()
+        return list(config.getDisplays())
+
+    def _get_default_display(self) -> str:
+        """Get default OCIO display."""
+        config = ocio.GetCurrentConfig()
+        return config.getDefaultDisplay()
+
+    def _get_views(self) -> list[str]:
+        """
+        Get all active OCIO views, given the current input color space.
+        """
+        config = ocio.GetCurrentConfig()
+        input_color_space = self.input_color_space()
+        if input_color_space:
+            return config.getViews(self.display(), input_color_space)
+        else:
+            return config.getViews(self.display())
+
+    def _get_default_view(self) -> str:
+        """
+        Get default OCIO view, given the current display and input
+        color space.
+        """
+        config = ocio.GetCurrentConfig()
+        input_color_space = None
+        if input_color_space:
+            return config.getDefaultView(self.display(), input_color_space)
+        else:
+            return config.getDefaultView(self.display())
+
+    def _on_mode_changed(self) -> None:
+        """Called when the application mode changes."""
+        mode = OCIOViewMode.current_mode()
+
+        if mode == OCIOViewMode.Preview:
+            self.mode_stack.setCurrentWidget(self.display_view_frame)
+        else:  # Edit
+            self.mode_stack.setCurrentWidget(self.tf_frame)
+
+        self.output_tf_direction_label.setVisible(mode == OCIOViewMode.Edit)
+        self.update(force=True)
 
     @QtCore.Slot(Path, int, int)
     def _on_image_loaded(
@@ -499,5 +838,118 @@ class BaseImageViewer(QtWidgets.QWidget):
         self.image_plane.update_gamma(value)
 
     @QtCore.Slot(int)
+    def _on_display_changed(self, index: int) -> None:
+        """Called when the display changes."""
+        with SignalsBlocked(self.view_box):
+            self.view_box.update_items()
+        self._on_view_changed(0)
+
+    @QtCore.Slot(int)
+    def _on_view_changed(self, index: int) -> None:
+        """Called when the view changes."""
+        self.update(update_items=False)
+
+    @QtCore.Slot(int)
     def _on_sample_precision_changed(self, value: float) -> None:
         self._sample_format = f"{{v:.{value}f}}"
+
+    def _on_transform_menu_changed(
+        self, menu_items: list[tuple[int, str, QtGui.QIcon]]
+    ) -> None:
+        """
+        Called to refresh transform menu items, and either reselect the
+        existing subscription, or deselect any subscription.
+        """
+        target_index = -1
+        current_slot = -1
+        if self.tf_box.count():
+            current_slot = self.tf_box.currentData(role=self.ROLE_SLOT)
+
+        with SignalsBlocked(self.tf_box):
+            self.tf_box.clear()
+
+            # The first item is always no transform
+            self.tf_box.addItem(self.PASSTHROUGH)
+            self.tf_box.setItemData(0, -1, role=self.ROLE_SLOT)
+
+            for i, (
+                slot,
+                item_label,
+                item_type,
+                item_name,
+                slot_icon,
+            ) in enumerate(menu_items):
+                index = i + 1
+                self.tf_box.addItem(slot_icon, item_label)
+                self.tf_box.setItemData(index, slot, role=self.ROLE_SLOT)
+                self.tf_box.setItemData(
+                    index, item_type, role=self.ROLE_ITEM_TYPE
+                )
+                self.tf_box.setItemData(
+                    index, item_name, role=self.ROLE_ITEM_NAME
+                )
+                if slot == current_slot:
+                    target_index = index  # Offset for "Passthrough" item
+
+            # Restore previous item?
+            if target_index != -1:
+                self.tf_box.setCurrentIndex(target_index)
+
+        # Switch to "Passthrough" if previous slot not found
+        if target_index == -1 and self.tf_box.count():
+            with SignalsBlocked(self.tf_box):
+                self.tf_box.setCurrentIndex(0)
+
+            # Force update transform
+            self._on_transform_changed(0)
+
+    def _on_transform_subscription_init(self, slot: int) -> None:
+        """
+        If this viewer is not subscribed to a specific transform
+        subscription slot, subscribe to the first slot to receive a
+        transform subscription.
+
+        :param slot: Transform subscription slot
+        """
+        if self._tf_subscription_slot == -1:
+            index = self.tf_box.findData(slot, role=self.ROLE_SLOT)
+            if index != -1:
+                self.tf_box.setCurrentIndex(index)
+
+    @QtCore.Slot(int)
+    def _on_transform_changed(self, index: int) -> None:
+        if index == 0:
+            TransformManager.unsubscribe_from_all_transforms(
+                self.set_transform
+            )
+            self.clear_transform()
+        else:
+            self._tf_subscription_slot = self.tf_box.currentData(
+                role=self.ROLE_SLOT
+            )
+            TransformManager.subscribe_to_transforms_at(
+                self._tf_subscription_slot, self.set_transform
+            )
+
+    @QtCore.Slot(int)
+    def _on_tf_subscription_requested(self, slot: int) -> None:
+        # If the requested slot does not have a subscription, "Passthrough" will
+        # be selected.
+        self.tf_box.setCurrentIndex(
+            max(0, self.tf_box.findData(slot, role=self.ROLE_SLOT))
+        )
+
+    @QtCore.Slot(bool)
+    def _on_inverse_check_clicked(self, checked: bool) -> None:
+        self.set_transform(
+            self._tf_subscription_slot, self._tf_fwd, self._tf_inv
+        )
+        if self.tf_direction_button.isChecked():
+            self.tf_direction_button.setIcon(self._tf_direction_inverse_icon)
+            self.tf_direction_button.setToolTip("Transform direction: Inverse")
+            # Use 'minus' character to match the width of "+"
+            self.output_tf_direction_label.setText("\u2212")
+        else:
+            self.tf_direction_button.setIcon(self._tf_direction_forward_icon)
+            self.tf_direction_button.setToolTip("Transform direction: Forward")
+            self.output_tf_direction_label.setText("+")
