+++ src/apps/ocioview/ocioview/widgets/structure.py
@@ -18,11 +18,14 @@ class DockTitleBar(QtWidgets.QFrame):
         self,
         title: str,
         icon: QtGui.QIcon,
+        widget: Optional[QtWidgets.QWidget] = None,
         parent: Optional[QtCore.QObject] = None,
     ):
         """
         :param title: Title text
         :param icon: Dock icon
+        :param widget: Optional widget to display opposite the title
+            and icon.
         """
         super().__init__(parent=parent)
 
@@ -36,6 +39,7 @@ class DockTitleBar(QtWidgets.QFrame):
         self.icon = QtWidgets.QLabel()
         self.icon.setPixmap(icon.pixmap(ICON_SIZE_ITEM))
         self.title = QtWidgets.QLabel(title)
+        self.widget = widget
 
         # Layout
         inner_layout = QtWidgets.QHBoxLayout()
@@ -44,6 +48,8 @@ class DockTitleBar(QtWidgets.QFrame):
         inner_layout.addWidget(self.icon)
         inner_layout.addWidget(self.title)
         inner_layout.addStretch()
+        if widget is not None:
+            inner_layout.addWidget(self.widget)
 
         inner_frame = QtWidgets.QFrame()
         inner_frame.setFrameShape(QtWidgets.QFrame.StyledPanel)
@@ -71,14 +77,17 @@ class TabbedDockWidget(QtWidgets.QDockWidget):
         self,
         title: str,
         icon: QtGui.QIcon,
+        corner_widget: Optional[QtWidgets.QWidget] = None,
         parent: Optional[QtCore.QObject] = None,
     ):
         """
         :param title: Title text
         :param icon: Dock icon
+        :param corner_widget: Optional widget to place on the right
+            side of the dock title bar.
         """
         super().__init__(parent=parent)
-        self.setTitleBarWidget(DockTitleBar(title, icon))
+        self.setTitleBarWidget(DockTitleBar(title, icon, widget=corner_widget))
         self.setFeatures(
             QtWidgets.QDockWidget.DockWidgetMovable
             | QtWidgets.QDockWidget.DockWidgetFloatable
